requirements.txt
# Core LLM - Official Gemini SDK (google-generativeai is legacy)
google-genai>=1.68.0

# LangChain ecosystem
langchain>=1.2.13
langchain-google-genai>=2.0.0
langgraph>=1.1.3

# Telegram Bot (fully async)
python-telegram-bot>=22.0

# Embeddings
sentence-transformers>=3.0.0

# Vector Store
sqlite-vec>=0.1.7

# Utilities
python-dotenv>=1.0.0
numpy>=1.26.0
